perf(code): 代码上传
暗黑模式修正

+++ web/static/js/app.js
@@ -455,19 +455,27 @@ new Vue({
             });
         },
         removeFile(fileId) {
-            fetch(`/api/files/${fileId}`, {
-                method: 'DELETE'
-            }).then(res => res.json())
-                .then(data => {
-                    if (data.success) {
-                        this.$message.success('删除成功！');
-                        this.loadFileList();
-                    } else {
-                        this.$message.error(data.message || '删除失败');
-                    }
-                }).catch(() => {
-                this.$message.error('删除失败');
-            })
+            this.$confirm('确定要删除这个文件吗?', '提示', {
+                confirmButtonText: '确定',
+                cancelButtonText: '取消',
+                type: 'warning'
+            }).then(() => {
+                fetch(`/api/files/${fileId}`, {
+                    method: 'DELETE'
+                }).then(res => res.json())
+                    .then(data => {
+                        if (data.success) {
+                            this.$message.success('删除成功！');
+                            this.loadFileList();
+                        } else {
+                            this.$message.error(data.message || '删除失败');
+                        }
+                    }).catch(() => {
+                    this.$message.error('删除失败');
+                })
+            }).catch(() => {
+                this.$message.info('已取消删除');
+            });
         },
 
         getLevelColor(level){
